Move interface to a different package
