+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/IntegrationTestsSetup.java
@@ -7,8 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerInitializedEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.cmc.claimstore.BaseSaveTest;
-import uk.gov.hmcts.cmc.claimstore.aat.TestsSetup;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 
 @Component
