+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/RestTestClient.java
@@ -5,7 +5,6 @@ import io.restassured.response.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.cmc.claimstore.aat.TestsSetup;
 import uk.gov.hmcts.cmc.claimstore.processors.JsonMapper;
 import uk.gov.hmcts.cmc.domain.models.ClaimData;
